[AMDGPU] Keep loop-interleave pass private to jit_amdgpu.cpp
Move AMDGPUSetLoopInterleavePass out of the shared llvm_context_pass.h
and into an anonymous namespace in jit_amdgpu.cpp (its only user), under
the existing QD_WITH_AMDGPU guard, with the LoopInfo/LoopPassManager
includes local to that translation unit. This keeps the common LLVM
header free of backend-specific code so CPU/CUDA include sites no longer
compile or depend on it. Pure relocation: pass logic and pipeline
registration are unchanged.
