+++ quadrants/runtime/llvm/llvm_context_pass.h
@@ -298,63 +298,5 @@ struct AMDGPUConvertFuncParamAddressSpacePass : public ModulePass {
 
 #endif
 
-// New-PM loop pass: annotates innermost loops with llvm.loop.interleave.count
-// IR metadata so the loop vectorizer uses the requested interleave factor.
-// Scoped to the compilation pipeline where it is registered; does not touch
-// any process-wide LLVM command-line state.
-struct AMDGPUSetLoopInterleavePass
-    : public llvm::PassInfoMixin<AMDGPUSetLoopInterleavePass> {
-  unsigned count_;
-  explicit AMDGPUSetLoopInterleavePass(unsigned count) : count_(count) {}
-
-  llvm::PreservedAnalyses run(llvm::Loop &L,
-                               llvm::LoopAnalysisManager &,
-                               llvm::LoopStandardAnalysisResults &,
-                               llvm::LPMUpdater &) {
-    // Only annotate innermost loops
-    if (!L.getSubLoops().empty())
-      return llvm::PreservedAnalyses::all();
-
-    llvm::LLVMContext &ctx = L.getHeader()->getContext();
-    llvm::MDNode *existing_id = L.getLoopID();
-
-    // Skip if interleave count is already specified
-    if (existing_id) {
-      for (unsigned i = 1, e = existing_id->getNumOperands(); i < e; ++i) {
-        if (auto *node = llvm::dyn_cast<llvm::MDNode>(existing_id->getOperand(i))) {
-          if (node->getNumOperands() > 0) {
-            if (auto *s = llvm::dyn_cast<llvm::MDString>(node->getOperand(0))) {
-              if (s->getString() == "llvm.loop.interleave.count")
-                return llvm::PreservedAnalyses::all();
-            }
-          }
-        }
-      }
-    }
-
-    // Build new interleave hint node
-    llvm::MDNode *hint = llvm::MDNode::get(ctx, {
-        llvm::MDString::get(ctx, "llvm.loop.interleave.count"),
-        llvm::ConstantAsMetadata::get(
-            llvm::ConstantInt::get(llvm::Type::getInt32Ty(ctx), count_))
-    });
-
-    // Assemble updated loop metadata: [self-ref, ...existing..., hint]
-    llvm::SmallVector<llvm::Metadata *, 4> ops;
-    ops.push_back(nullptr); // placeholder for self-reference
-    if (existing_id) {
-      for (unsigned i = 1, e = existing_id->getNumOperands(); i < e; ++i)
-        ops.push_back(existing_id->getOperand(i).get());
-    }
-    ops.push_back(hint);
-
-    llvm::MDNode *new_id = llvm::MDNode::get(ctx, ops);
-    new_id->replaceOperandWith(0, new_id); // fix self-reference
-    L.setLoopID(new_id);
-
-    return llvm::PreservedAnalyses::none();
-  }
-};
-
 }  // namespace lang
 }  // namespace quadrants
